feat: collect per-test logs of root logger to file
Create a new logging handler for each test to add to the root logger, so
all the logs of that logger are stored to a file; this file is then
added to the artifacts of each test, if it was created.

Make sure to properly remove the added handlers, so each handler does
not interfere with other tests.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>