refactor: move "subman" implementation to common function
There will be soon another version of the "subman" fixture with a
different scope; since pytest requires precise scopes for each fixture,
then there will be different fixtures for other scopes.

To help with that, move the implementation of the "subman" fixture to an
helper function, leaving only a simple call to that function in
"subman".

There should be no behaviour changes.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>