feat: add new "subman_global" fixture
Add a new "subman_global" fixture to act on subscription-manager as a
single interaction for the whole session. This is needed because the
scope of each pytest fixture prevents the fixture to be used in other
scopes, and "subman" has a "function" scope.

Other than the scope, there should be no behaviour difference between
"subman" and "subman_global".

Signed-off-by: Pino Toscano <ptoscano@redhat.com>