feat: save/restore also insights-client-payload.log for insights-client
insights-client-payload.log is used instead of insights-client.log in
case the --payload option used. Hence, do the backup also of that log
file, restoring it when done, and archiving it if created during a test.

Also handle the .1/.2/.3 versions, to handle the old client/core
versions that did their own log rotation.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>