feat: Podman: improve requiring/checking for running container
In case a method required a running container, it used to check on its
own, and return something. While this works, it is not properly
communicating this situation to callers, and it requires manual checks
to be added in new methods.

To improve the situation, then:
- a new internal "ContainerNotRunningError" exception represents
  whether a container is not running
- a new internal "requires_running" decorator checks whether a container
  is running (i.e. it has an ID), raising "ContainerNotRunningError" if
  not
- "Podman.stop()" is changed to use the new "requires_running"
  decorator; because of this, the usage of "stop()" in "__exit__()" must
  properly keep ignoring this situation, to allow for the container to
  be stopped manually even before a context manager ends

All in all, this should be an internal refactor with no actual behaviour
changes for the plugin.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>