fix: InsightsClient: fix handling of bool+string config keys
Apparently there are configuration keys that may allow both boolean and
string values, and "cert_verify" is a notable one.

To handle this properly, add a new category of keys, and tweak the
reading of keys to fallback with a string-based get() in case the
(likely from getboolean()) reading fails with ValueError.

Add a couple of tests to verify that the right types are properly
handled.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>