feat: add internal helper to redact arguments
To avoid logging sensitive information (e.g. credentials), it is
necessary to redact the command line invocations done by the classes of
the tools.

To do this, introduce a new redact_arguments() helper that takes a list
of arguments to redact in form of "--foo", etc; this helper redacts
either
- the argument after each argument exactly as found in the redaction
  list; for example in ["--foo", "value"] "value" is redacted if "--foo"
  is specified to be redacted
- the argument itself if it is the start of a parameter with value; for
  example "--foo=value" is redacted if "--foo" is specified to be
  redacted

Signed-off-by: Pino Toscano <ptoscano@redhat.com>